Update to 1.0.3
Full Changelog:
- Change driver support to "config composer" to make it more easy to
handle the devices.
- Updated the z-wave driver to 1.0.8 which includes a few extra options
(save value on reboot when less then 30 minutes old) and adds tiny bit
of stability
- Made all coding more compliant with ES6
- Enhance a few learning mode teksts.
TSP01
TZ67 (-PLUS)
TZ65 (-PLUS)
- Improved/updated all icons that needed an update.
TSP01
TZ36
TZ65
TZ66
- Enhanced Basic_Set handeling, which should improve responsiveness to
normal switches, and should also fix the right switch not triggering
Do make sure homey's ID (probably 1) is in the correct assocation
groups, so only in 2 and 3
If you put it in group 1 it will not work as it is supposed to, group 4
is not needed for the workings in homey.